Moved all asserts to the BasePage
