+++ Biopoolsengardens/Biopoolsengardens/BasePage.cs
@@ -1,4 +1,5 @@
-﻿using OpenQA.Selenium;
+﻿using NUnit.Framework;
+using OpenQA.Selenium;
 using OpenQA.Selenium.Support.UI;
 using System;
 
@@ -37,10 +38,84 @@ namespace Biopoolsengardens
             Driver.Url = url;
         }
 
-    
+        public void AssertIsDisplayedGratisOfferLink(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertLink.Text);
+
+        }
+
+
+        public void MakeApointmentAssert(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertLink2.Text);
+
+        }
+
+        public void BioPoolsAssertLink(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertBioPoolsLink.Text);
+
+        }
+
+        public void SwimmingPondsAssertLink(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertSwimmingPondsLink.Text);
+
+        }
+
+        public void GardenAndNaturalPondsLink(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertGardenAndNaturalPondsLink.Text);
+
+        }
+
+        public void SwimmingPoolsLink(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertSwimmingPoolsLink.Text);
+
+        }
+
+        public void RealizationLink(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertRealizationLink.Text);
+
+        }
+
+        public void ContactLink(string expected)
+        {
+
+            Assert.AreEqual(expected, AssertContactLink.Text);
+
+        }
+
+
+        public IWebElement AssertLink => Driver.FindElement(By.Id("element-883"));
+
+        public IWebElement AssertLink2 => Driver.FindElement(By.Id("element-773"));
+
+        public IWebElement AssertBioPoolsLink => Driver.FindElement(By.LinkText("Biozwembaden"));
+
+        public IWebElement AssertSwimmingPondsLink => Driver.FindElement(By.LinkText("Zwemvijvers"));
+
+        public IWebElement AssertGardenAndNaturalPondsLink => Driver.FindElement(By.LinkText("Tuin- en natuurvijvers"));
+
+        public IWebElement AssertSwimmingPoolsLink => Driver.FindElement(By.LinkText("Zwembaden"));
+
+        public IWebElement AssertRealizationLink => Driver.FindElement(By.LinkText("Realisaties"));
+
+        public IWebElement AssertContactLink => Driver.FindElement(By.LinkText("Contact"));
+
+
 
     }
 
 
-   
+
 }
