+++ Biopoolsengardens/Biopoolsengardens/Pages/MakeApointmentPage/MakeAnAppointment.cs
@@ -202,7 +202,6 @@ namespace Biopoolsengardens.Pages
                 }
                 _driver.Quit();
 
-
             }
         }
